+++ packages/create-app/src/constants.ts
@@ -1,9 +1,9 @@
 import { fileURLToPath } from 'node:url';
-import { versions as localVersions } from '@/versions';
 import versionPkg from '../../create-app-versions/package.json';
 
 export const sourceDir = fileURLToPath(new URL(`../`, import.meta.url).href);
-export const cwd = process.cwd();
+
+export const isCI = Boolean(process.env.CI);
 
 export interface TemplateInfo {
   value:
@@ -13,7 +13,16 @@ export interface TemplateInfo {
     | 'react-router-spa'
     | 'tanstack-start';
   label: string;
+  appDir: string;
+  /**
+   * path to root provider, relative to `appDir``
+   */
+  rootProviderPath: string;
   hint?: string;
+  /**
+   * rename files when copying from template
+   */
+  rename?: (name: string) => string;
 }
 
 export const templates: TemplateInfo[] = [
@@ -21,27 +30,34 @@ export const templates: TemplateInfo[] = [
     value: '+next+fuma-docs-mdx',
     label: 'Next.js: Fumadocs MDX',
     hint: 'recommended',
+    appDir: '',
+    rootProviderPath: 'app/layout.tsx',
   },
   {
     value: 'waku',
     label: 'Waku: Fumadocs MDX',
+    appDir: 'src',
+    rootProviderPath: 'components/provider.tsx',
   },
   {
     value: 'react-router',
     label: 'React Router: Fumadocs MDX (not RSC)',
+    appDir: 'app',
+    rootProviderPath: 'root.tsx',
   },
   {
     value: 'react-router-spa',
     label: 'React Router SPA: Fumadocs MDX (not RSC)',
     hint: 'SPA mode allows you to host the site statically, compatible with a CDN.',
+    appDir: 'app',
+    rootProviderPath: 'root.tsx',
   },
   {
     value: 'tanstack-start',
     label: 'Tanstack Start: Fumadocs MDX (not RSC)',
+    appDir: 'src',
+    rootProviderPath: 'routes/__root.tsx',
   },
 ];
 
-export const depVersions = {
-  ...localVersions,
-  ...versionPkg.dependencies,
-};
+export const depVersions = versionPkg.dependencies;