+++ packages/create-app/src/transform/react-router.ts
@@ -8,25 +8,24 @@ import { getCodeValue } from '@/transform/shared';
 import SyntaxKind = ts.SyntaxKind;
 
 /**
- * remove items from `excluded` in the prerender function
+ * filter items in a specific array initializer in the prerender function
  */
-export function removeReactRouterPrerenderExclude(
+export function filterReactRouterPrerenderArray(
   sourceFile: SourceFile,
-  paths: string[],
+  array: 'paths' | 'excluded',
+  filter: (item: string) => boolean,
 ) {
   const methodBody = getPrerenderMethod(sourceFile)?.getBody();
   if (!methodBody) return;
 
   const initializer = methodBody
     .getDescendantsOfKind(SyntaxKind.VariableDeclaration)
-    .find((item) => item.getName() === 'excluded')
+    .find((item) => item.getName() === array)
     ?.getInitializerIfKind(SyntaxKind.ArrayLiteralExpression);
 
   if (!initializer) return;
   for (const element of initializer.getElements()) {
-    const value = getCodeValue(element.getText());
-
-    if (paths.includes(value)) {
+    if (!filter(getCodeValue(element.getText()))) {
       initializer.removeElement(element);
     }
   }
@@ -51,16 +50,22 @@ export function addReactRouterRoute(
  */
 export function filterReactRouterRoute(
   sourceFile: SourceFile,
-  filter: (item: { path: string }) => boolean,
+  filter: (item: { path: string; entry: string }) => boolean,
 ) {
   modifyReactRouterRoutes(sourceFile, (arr) => {
     for (const element of arr.getElements()) {
       if (
         !element.isKind(SyntaxKind.CallExpression) ||
         element.getFirstChildByKind(SyntaxKind.Identifier)?.getText() !==
-          'route' ||
+          'route'
+      )
+        continue;
+      const args = element.getArguments();
+
+      if (
         filter({
-          path: getCodeValue(element.getArguments()[0].getText()),
+          path: getCodeValue(args[0].getText()),
+          entry: getCodeValue(args[1].getText()),
         })
       )
         continue;