+++ packages/mdx/CHANGELOG.md
@@ -1,5 +1,20 @@
 # next-docs-mdx
 
+## 13.0.2
+
+### Patch Changes
+
+- ee4ad3d: Always format file paths into POSIX for Vite
+
+## 13.0.1
+
+### Patch Changes
+
+- 56332df: Support plugins in Webpack loader environment: now plugins can apply changes on MDX options too.
+- 91add4f: Plugin `json-schema`: support inserting `$schema` to JSON files
+- cffd4c2: Lazy update index files on Next.js
+  - fumadocs-core@16.0.3
+
 ## 13.0.0
 
 ### Major Changes